added commands to buttons
